+++ src/bot/MessageHandler.ts
@@ -1,26 +1,25 @@
 import { Message } from 'discord.js';
 import '../discord/Message';
 
-import config from '@config/config.json';
-
+import Config from '@config/Config';
 import DatabaseAdapter from '@util/db/DatabaseAdapter';
 import CommandCollection from './CommandCollection';
 
 export default class MessageHandler {
+  private readonly config: Config;
   private readonly db: DatabaseAdapter;
   private readonly commands: CommandCollection;
-  private readonly prefix: string;
 
-  constructor(commands: CommandCollection, db: DatabaseAdapter) {
+  constructor(config: Config, commands: CommandCollection, db: DatabaseAdapter) {
+    this.config = config;
     this.db = db;
     this.commands = commands;
-    this.prefix = config.prefix;
   }
 
   public handle(message: Message) {
     if (!this.isValidMessage(message)) return;
 
-    message.content = message.content.substring(this.prefix.length);
+    message.content = message.content.substring(this.config.prefix.length);
     const [command, ...params] = message.content.split(' ');
 
     this.commands.execute(command, params, message);
@@ -28,7 +27,7 @@ export default class MessageHandler {
 
   private isValidMessage(message: Message) {
     return !message.isDirectMessage() &&
-           message.hasPrefix(this.prefix) &&
+           message.hasPrefix(this.config.prefix) &&
            !this.db.ignoreList.exists(message.author.id);
   }
 }