+++ src/bot/commands/SoundsCommand.ts
@@ -1,19 +1,20 @@
 import { Message } from 'discord.js';
 
-import config from '@config/config.json';
-
 import ICommand from './base/ICommand';
 
+import Config from '@config/Config';
 import LocaleService from '@util/i18n/LocaleService';
 import SoundUtil from '@util/SoundUtil';
 import MessageChunker from './helpers/MessageChunker';
 
 export default class SoundsCommand implements ICommand {
   public readonly TRIGGERS = ['sounds'];
+  private readonly config: Config;
   private readonly localeService: LocaleService;
   private readonly chunker: MessageChunker;
 
-  constructor(localeService: LocaleService, chunker: MessageChunker) {
+  constructor(config: Config, localeService: LocaleService, chunker: MessageChunker) {
+    this.config = config;
     this.localeService = localeService;
     this.chunker = chunker;
   }
@@ -22,7 +23,7 @@ export default class SoundsCommand implements ICommand {
     const sounds = SoundUtil.getSounds();
 
     if (!sounds.length) {
-      message.author.send(this.localeService.t('sounds.notFound', { prefix: config.prefix }));
+      message.author.send(this.localeService.t('sounds.notFound', { prefix: this.config.prefix }));
       return;
     }