+++ src/bot/commands/WelcomeCommand.ts
@@ -1,20 +1,21 @@
 import { Message } from 'discord.js';
 
-import config from '@config/config.json';
-
 import ICommand from './base/ICommand';
 
+import Config from '@config/Config';
 import LocaleService from '@util/i18n/LocaleService';
 
 export default class WelcomeCommand implements ICommand {
   public readonly TRIGGERS = ['welcome'];
+  private readonly config: Config;
   private readonly localeService: LocaleService;
 
-  constructor(localeService: LocaleService) {
+  constructor(config: Config, localeService: LocaleService) {
+    this.config = config;
     this.localeService = localeService;
   }
 
   public run(message: Message, _: Array<string>) {
-    message.channel.send(this.localeService.t('welcome', { prefix: config.prefix }));
+    message.channel.send(this.localeService.t('welcome', { prefix: this.config.prefix }));
   }
 }