+++ src/util/queue/SoundQueue.ts
@@ -1,17 +1,18 @@
 import { VoiceConnection } from 'discord.js';
 
-import config from '@config/config.json';
-
+import Config from '@config/Config';
 import DatabaseAdapter from '@util/db/DatabaseAdapter';
 import SoundUtil from '@util/SoundUtil';
 import QueueItem from './QueueItem';
 
 export default class SoundQueue {
+  private readonly config: Config;
   private readonly db: DatabaseAdapter;
   private queue: Array<QueueItem>;
   private currentSound: QueueItem | null;
 
-  constructor(db: DatabaseAdapter) {
+  constructor(config: Config, db: DatabaseAdapter) {
+    this.config = config;
     this.db = db;
     this.queue = [];
     this.currentSound = null;
@@ -47,19 +48,20 @@ export default class SoundQueue {
 
   private deafen(connection: VoiceConnection) {
     // Can only deafen when in a channel, therefore need connection
-    connection.channel.guild.me.setDeaf(config.deafen);
+    connection.channel.guild.me.setDeaf(this.config.deafen);
     return Promise.resolve(connection);
   }
 
   private playSound(connection: VoiceConnection, name: string) {
     return new Promise(resolve =>
-      connection.playFile(name, { volume: config.volume }).on('end', () => resolve(connection))
+      connection.playFile(name, { volume: this.config.volume })
+                .on('end', () => resolve(connection))
     );
   }
 
   private onFinishedPlayingSound(connection: VoiceConnection) {
     this.db.sounds.incrementCount(this.currentSound!.name);
-    if (config.deleteMessages) this.currentSound!.message.delete();
+    if (this.config.deleteMessages) this.currentSound!.message.delete();
 
     if (!this.isEmpty()) {
       this.playNext();
@@ -67,7 +69,7 @@ export default class SoundQueue {
     }
 
     this.currentSound = null;
-    if (!config.stayInChannel) connection.disconnect();
+    if (!this.config.stayInChannel) connection.disconnect();
   }
 
   private isEmpty() {